feat: Assured method 분리
